requirements.txt
apscheduler==3.6.0
asn1crypto==0.24.0
AWSIoTPythonSDK==1.4.4
bcrypt==3.1.6
boto3==1.9.121
botocore==1.12.121
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
docutils==0.14
idna==2.8
jmespath==0.9.4
paramiko==2.4.2
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
python-dateutil==2.8.0
python-nmap==0.6.1
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
tzlocal==1.5.1
urllib3==1.24.1
